Changed function of "pokemonList" to "trait". Updated formatting and syntaxing
List was only displaying a handful of pokemon. Reason was because of an undeclared function. Changed the function to be defined as "trait" and updated the rest. All pokemon now show up!

Didn't like the way the "That's a big Pokemon!" text was displayed so I changed the formatting with <br> and added parentheses to separate height.